+++ src/audio/microphone.h
@@ -1,7 +1,7 @@
+#include "util.h"
 #include "types.h"
 #include <AL/al.h>
 #include <AL/alc.h>
-#include <stdbool.h>
 
 #pragma once
 
@@ -12,20 +12,20 @@ typedef struct Microphone {
   ALCdevice* device;
   const char* name;
   bool isRecording;
-  int sampleRate;
-  int bitDepth;
-  int channelCount;
+  u32 sampleRate;
+  u32 bitDepth;
+  u32 channelCount;
 } Microphone;
 
-Microphone* lovrMicrophoneInit(Microphone* microphone, const char* name, int samples, int sampleRate, int bitDepth, int channelCount);
+Microphone* lovrMicrophoneInit(Microphone* microphone, const char* name, usize samples, u32 sampleRate, u32 bitDepth, u32 channelCount);
 #define lovrMicrophoneCreate(...) lovrMicrophoneInit(lovrAlloc(Microphone), __VA_ARGS__)
 void lovrMicrophoneDestroy(void* ref);
-int lovrMicrophoneGetBitDepth(Microphone* microphone);
-int lovrMicrophoneGetChannelCount(Microphone* microphone);
+u32 lovrMicrophoneGetBitDepth(Microphone* microphone);
+u32 lovrMicrophoneGetChannelCount(Microphone* microphone);
 struct SoundData* lovrMicrophoneGetData(Microphone* microphone);
 const char* lovrMicrophoneGetName(Microphone* microphone);
-int lovrMicrophoneGetSampleCount(Microphone* microphone);
-int lovrMicrophoneGetSampleRate(Microphone* microphone);
+usize lovrMicrophoneGetSampleCount(Microphone* microphone);
+u32 lovrMicrophoneGetSampleRate(Microphone* microphone);
 bool lovrMicrophoneIsRecording(Microphone* microphone);
 void lovrMicrophoneStartRecording(Microphone* microphone);
 void lovrMicrophoneStopRecording(Microphone* microphone);
